+++ packages/cli/__tests__/zosjobs/__integration__/view/cli.zos-jobs.view.job-status-by-jobid.integration.test.ts
@@ -51,7 +51,12 @@ describe("zos-jobs view job-status-by-jobid command", () => {
                 TEST_ENVIRONMENT);
             expect(response.stdout.toString()).toBe("");
             expect(response.status).toBe(1);
-            expect(response.stderr.toString()).toMatchSnapshot();
+            expect(response.stderr.toString()).toContain('Unknown argument: blah');
+            expect(response.stderr.toString()).toContain('Command failed due to improper syntax');
+            expect(response.stderr.toString()).toContain('Command entered: "zos-jobs view job-status-by-jobid JOB123 blah ' +
+                '--host fakehost --user fakeuser --password fakepass"');
+            expect(response.stderr.toString()).toContain('Available commands are "all-spool-content, job-status-by-jobid, spool-file-by-id".');
+            expect(response.stderr.toString()).toContain('Use "zowe zos-jobs view --help" to view groups, commands, and options.');
         });
     });